+++ docs/changes.rst
@@ -7,6 +7,7 @@ Release 1.1.0
 
 * Python 3 compatibility
 * New module: :mod:`pytiger.utils.plugins`
+* New decorator: :class:`pytiger.utils.decorators.singleton`
 
 Release 1.0.0
 =============